move the fullNotes up a level
