+++ client/components/octave.tsx
@@ -2,7 +2,7 @@
 // A container grouping the keys together.
 // Each octave within keyboard-full will have 12 keys while the octave in keyboard-selected
 // will have 13 keys (12 standard keys plus the starting key of the next octave).
-// Each fullNote in the fullNotes array begins with a note and ends with an octave number, for example 'C3'
+// Each fullNote in the fullNotes array begins with a note and ends with an octave number, for example 'C4'
 
 import Key from '@/components/key';
 import type { FullNote } from '@/types/keyboard-option-types';
